requirements.txt
alembic==0.9.9
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookies==2.2.1
coverage==4.5.1
Flask==1.0.2
Flask-Dance==0.14.0
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.8.1
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.1.0
oauth2==1.9.0.post1
oauthlib==2.0.7
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pytest==3.5.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
requests==2.18.4
requests-oauthlib==0.8.0
responses==0.9.0
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.1
